+++ src/Account.php
@@ -34,72 +34,274 @@
  */
 
 /**
- * Data Transfer Object representing a Nosto account.
+ * Account DTO (Data Transfer Object).
  */
-class NostoAccount
+class NostoAccount implements \NostoAccountMetaInterface
 {
     /**
-     * @var string the name of the Nosto account.
+     * @var string the store name.
      */
-    private $name;
+    protected $_title;
+
+    /**
+     * @var string the account name.
+     */
+    protected $_name;
+
+    /**
+     * @var string the store front end url.
+     */
+    protected $_frontPageUrl;
+
+    /**
+     * @var \NostoCurrencyCode the store currency ISO (ISO 4217) code.
+     */
+    protected $_currency;
+
+    /**
+     * @var \NostoLanguageCode the store language ISO (ISO 639-1) code.
+     */
+    protected $_language;
+
+    /**
+     * @var \NostoLanguageCode the owner language ISO (ISO 639-1) code.
+     */
+    protected $_ownerLanguage;
+
+    /**
+     * @var \NostoOwner the account owner meta model.
+     */
+    protected $_owner;
+
+    /**
+     * @var \NostoAccountMetaBillingInterface the billing meta model.
+     */
+    protected $_billing;
+
+    /**
+     * @var \NostoCurrency[] list of supported currencies by the store.
+     */
+    protected $_currencies = array();
+
+    /**
+     * @var string the default price variation ID if using multiple currencies.
+     */
+    protected $_defaultPriceVariationId;
+
+    /**
+     * @var bool if the store uses exchange rates to manage multiple currencies.
+     */
+    protected $_useCurrencyExchangeRates = false;
+
+    /**
+     * @var string the API token used to identify an account creation.
+     */
+    protected $_signUpApiToken = 'YBDKYwSqTCzSsU8Bwbg4im2pkHMcgTy9cCX7vevjJwON1UISJIwXOLMM0a8nZY7h';
 
     /**
      * @var NostoApiToken[] the Nosto API tokens associated with this account.
      */
-    protected $tokens = array();
+    protected $_tokens = array();
 
     /**
-     * Constructor.
-     * Create a new account object with given name.
-     *
+     * NostoAccount constructor.
      * @param $name
-     *
-     * @throws NostoInvalidArgumentException
      */
-    public function __construct($name)
+    public function __construct($name = '')
     {
-        if (!is_string($name) || empty($name)) {
-            throw new NostoInvalidArgumentException(sprintf(
-                '%s.name (%s) must be a non-empty string value.',
-                __CLASS__,
-                $name
-            ));
-        }
-
-        $this->name = (string)$name;
+        $this->_name = $name;
     }
 
     /**
-     * Gets the account name.
-     *
-     * @return string the account name.
+     * @inheritdoc
+     */
+    public function getTitle()
+    {
+        return $this->_title;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function setTitle($title)
+    {
+        $this->_title = $title;
+    }
+
+    /**
+     * @inheritdoc
      */
     public function getName()
     {
-        return $this->name;
+        return $this->_name;
     }
 
     /**
-     * Returns the accounts API tokens.
-     *
-     * @return NostoApiToken[] the tokens.
+     * @inheritdoc
      */
-    public function getTokens()
+    public function setName($name)
     {
-        return $this->tokens;
+        $this->_name = $name;
     }
 
     /**
-     * Checks if this account is the same as the given account.
-     * They are considered equal if their name property match. The tokens are not relevant in the comparison,
-     * as they are not required by the account upon creation.
-     *
-     * @param NostoAccount $account the account to check.
-     * @return bool true if equals.
+     * @inheritdoc
      */
-    public function equals(NostoAccount $account)
+    public function getPlatform()
     {
-        return $account->getName() === $this->getName();
+        return 'magento';
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getFrontPageUrl()
+    {
+        return $this->_frontPageUrl;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function setFrontPageUrl($frontPageUrl)
+    {
+        $this->_frontPageUrl = $frontPageUrl;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getCurrency()
+    {
+        return $this->_currency;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function setCurrency(\NostoCurrencyCode $currency)
+    {
+        $this->_currency = $currency;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getLanguage()
+    {
+        return $this->_language;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function setLanguage(\NostoLanguageCode $language)
+    {
+        $this->_language = $language;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getOwnerLanguage()
+    {
+        return $this->_ownerLanguage;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function setOwnerLanguage(\NostoLanguageCode $ownerLanguage)
+    {
+        $this->_ownerLanguage = $ownerLanguage;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getOwner()
+    {
+        return $this->_owner;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function setOwner(\NostoAccountMetaOwnerInterface $owner)
+    {
+        $this->_owner = $owner;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getBillingDetails()
+    {
+        return $this->_billing;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getCurrencies()
+    {
+        return $this->_currencies;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function setCurrencies(array $currencies)
+    {
+        $this->_currencies = $currencies;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getDefaultPriceVariationId()
+    {
+        return $this->_defaultPriceVariationId;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getUseCurrencyExchangeRates()
+    {
+        return $this->_useCurrencyExchangeRates;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function setUseCurrencyExchangeRates($useCurrencyExchangeRates)
+    {
+        $this->_useCurrencyExchangeRates = $useCurrencyExchangeRates;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getSignUpApiToken()
+    {
+        return $this->_signUpApiToken;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getPartnerCode()
+    {
+        return null;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function setBilling(\NostoAccountMetaBillingInterface $billing)
+    {
+        $this->_billing = $billing;
     }
 
     /**
@@ -114,17 +316,14 @@ class NostoAccount
     }
 
     /**
-     * Returns a list of API token names that are present for the account.
-     * The API tokens act as scopes when doing OAuth requests to Nosto.
-     *
-     * @return array the list of names.
+     * @inheritdoc
      */
     public function getMissingScopes()
     {
         $allTokens = NostoApiToken::getApiTokenNames();
         $foundTokens = array();
         foreach ($allTokens as $tokenName) {
-            foreach ($this->tokens as $token) {
+            foreach ($this->_tokens as $token) {
                 if ($token->getName() === $tokenName) {
                     $foundTokens[] = $tokenName;
                     break;
@@ -135,28 +334,39 @@ class NostoAccount
     }
 
     /**
-     * Adds an API token to the account.
-     *
-     * @param NostoApiToken $token the token.
+     * @inheritdoc
      */
     public function addApiToken(NostoApiToken $token)
     {
-        $this->tokens[] = $token;
+        $this->_tokens[] = $token;
     }
 
     /**
-     * Gets an api token associated with this account by it's name , e.g. "sso".
-     *
-     * @param string $name the api token name.
-     * @return NostoApiToken|null the token or null if not found.
+     * @inheritdoc
      */
     public function getApiToken($name)
     {
-        foreach ($this->tokens as $token) {
+        foreach ($this->_tokens as $token) {
             if ($token->getName() === $name) {
                 return $token;
             }
         }
         return null;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function equals(NostoAccount $account)
+    {
+        return $account->getName() === $this->getName();
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getTokens()
+    {
+        return $this->_tokens;
     }
 }