+++ src/Billing.php
@@ -34,57 +34,28 @@
  */
 
 /**
- * Interface for the OAuth2 client.
- * The client implements the "Authorization Code" OAuth2 spec.
- * @see https://tools.ietf.org/html/rfc6749
+ * Billing DTO (Data Transfer Object).
  */
-interface NostoOauthClientMetaInterface
+class NostoBilling implements \NostoAccountMetaBillingInterface
 {
     /**
-     * The OAuth2 client ID.
-     * This will be a platform specific ID that Nosto will issue.
-     *
-     * @return string the client id.
+     * @var \NostoCountryCode country ISO (ISO 3166-1 alpha-2) code for billing details.
      */
-    public function getClientId();
+    protected $_country;
 
     /**
-     * The OAuth2 client secret.
-     * This will be a platform specific secret that Nosto will issue.
-     *
-     * @return string the client secret.
+     * @inheritdoc
      */
-    public function getClientSecret();
+    public function getCountry()
+    {
+        return $this->_country;
+    }
 
     /**
-     * The OAuth2 redirect url to where the OAuth2 server should redirect the user after authorizing the application to
-     * act on the users behalf.
-     * This url must by publicly accessible and the domain must match the one defined for the Nosto account.
-     *
-     * @return string the url.
+     * @inheritdoc
      */
-    public function getRedirectUrl();
-
-    /**
-     * The scopes for the OAuth2 request.
-     * These are used to request specific API tokens from Nosto and should almost always be the ones defined in
-     * NostoApiToken::$tokenNames.
-     *
-     * @return array the scopes.
-     */
-    public function getScopes();
-
-    /**
-     * The 2-letter ISO code (ISO 639-1) for the language the OAuth2 server uses for UI localization.
-     *
-     * @return NostoLanguageCode the language code.
-     */
-    public function getLanguage();
-
-    /**
-     * The Nosto account if we are to sync account details from Nosto.
-     *
-     * @return NostoAccountMetaInterface the account.
-     */
-    public function getAccount();
+    public function setCountry(\NostoCountryCode $country)
+    {
+        $this->_country = $country;
+    }
 }