+++ src/Oauth.php
@@ -34,57 +34,109 @@
  */
 
 /**
- * Interface for the OAuth2 client.
- * The client implements the "Authorization Code" OAuth2 spec.
- * @see https://tools.ietf.org/html/rfc6749
+ * Oauth DTO (Data Transfer Object).
  */
-interface NostoOauthClientMetaInterface
+class NostoOauth implements \NostoOauthClientMetaInterface
 {
     /**
-     * The OAuth2 client ID.
-     * This will be a platform specific ID that Nosto will issue.
-     *
-     * @return string the client id.
+     * @var array The scopes for the OAuth2 request.
      */
-    public function getClientId();
+    protected $_scopes = array();
 
     /**
-     * The OAuth2 client secret.
-     * This will be a platform specific secret that Nosto will issue.
-     *
-     * @return string the client secret.
+     * @var string the url where the oauth2 server should redirect after
+     * authorization is done.
      */
-    public function getClientSecret();
+    protected $_redirectUrl;
 
     /**
-     * The OAuth2 redirect url to where the OAuth2 server should redirect the user after authorizing the application to
-     * act on the users behalf.
-     * This url must by publicly accessible and the domain must match the one defined for the Nosto account.
-     *
-     * @return string the url.
+     * @var \NostoLanguageCode the 2-letter ISO code (ISO 639-1) for localization
+     * on oauth2 server.
      */
-    public function getRedirectUrl();
+    protected $_language;
 
     /**
-     * The scopes for the OAuth2 request.
-     * These are used to request specific API tokens from Nosto and should almost always be the ones defined in
-     * NostoApiToken::$tokenNames.
-     *
-     * @return array the scopes.
+     * @var \NostoAccount|null account if OAuth is to sync details.
      */
-    public function getScopes();
+    protected $_account;
 
     /**
-     * The 2-letter ISO code (ISO 639-1) for the language the OAuth2 server uses for UI localization.
-     *
-     * @return NostoLanguageCode the language code.
+     * @inheritdoc
      */
-    public function getLanguage();
+    public function getClientId()
+    {
+        return 'magento';
+    }
 
     /**
-     * The Nosto account if we are to sync account details from Nosto.
-     *
-     * @return NostoAccountMetaInterface the account.
+     * @inheritdoc
      */
-    public function getAccount();
+    public function getClientSecret()
+    {
+        return 'magento';
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getScopes()
+    {
+        return $this->_scopes;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function setScopes(array $scopes)
+    {
+        $this->_scopes = $scopes;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getRedirectUrl()
+    {
+        return $this->_redirectUrl;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function setRedirectUrl($redirectUrl)
+    {
+        $this->_redirectUrl = $redirectUrl;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getLanguage()
+    {
+        return $this->_language;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function setLanguage(\NostoLanguageCode $language)
+    {
+        $this->_language = $language;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getAccount()
+    {
+        return $this->_account;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function setAccount(\NostoAccountMetaInterface $account)
+    {
+        $this->_account = $account;
+    }
 }