+++ src/Owner.php
@@ -34,57 +34,70 @@
  */
 
 /**
- * Interface for the OAuth2 client.
- * The client implements the "Authorization Code" OAuth2 spec.
- * @see https://tools.ietf.org/html/rfc6749
+ * Owner DTO (Data Transfer Object).
  */
-interface NostoOauthClientMetaInterface
+class NostoOwner implements \NostoAccountMetaOwnerInterface
 {
     /**
-     * The OAuth2 client ID.
-     * This will be a platform specific ID that Nosto will issue.
-     *
-     * @return string the client id.
+     * @var string the account owner first name.
      */
-    public function getClientId();
+    protected $_firstName;
 
     /**
-     * The OAuth2 client secret.
-     * This will be a platform specific secret that Nosto will issue.
-     *
-     * @return string the client secret.
+     * @var string the account owner last name.
      */
-    public function getClientSecret();
+    protected $_lastName;
 
     /**
-     * The OAuth2 redirect url to where the OAuth2 server should redirect the user after authorizing the application to
-     * act on the users behalf.
-     * This url must by publicly accessible and the domain must match the one defined for the Nosto account.
-     *
-     * @return string the url.
+     * @var string the account owner email address.
      */
-    public function getRedirectUrl();
+    protected $_email;
 
     /**
-     * The scopes for the OAuth2 request.
-     * These are used to request specific API tokens from Nosto and should almost always be the ones defined in
-     * NostoApiToken::$tokenNames.
-     *
-     * @return array the scopes.
+     * @inheritdoc
      */
-    public function getScopes();
+    public function getFirstName()
+    {
+        return $this->_firstName;
+    }
 
     /**
-     * The 2-letter ISO code (ISO 639-1) for the language the OAuth2 server uses for UI localization.
-     *
-     * @return NostoLanguageCode the language code.
+     * @inheritdoc
      */
-    public function getLanguage();
+    public function setFirstName($firstName)
+    {
+        $this->_firstName = $firstName;
+    }
 
     /**
-     * The Nosto account if we are to sync account details from Nosto.
-     *
-     * @return NostoAccountMetaInterface the account.
+     * @inheritdoc
      */
-    public function getAccount();
+    public function getLastName()
+    {
+        return $this->_lastName;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function setLastName($lastName)
+    {
+        $this->_lastName = $lastName;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getEmail()
+    {
+        return $this->_email;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function setEmail($email)
+    {
+        $this->_email = $email;
+    }
 }