+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.4.1]
+### Changed
+- Impelement more classes in SDK
+- Improve multi currency handling
+
 ## [2.4.0]
 ### Changed
 - Loosen up the data validation and handle incorrect data types